added checks in authentication

+++ backend/src/utils/authentication.py
@@ -236,19 +236,23 @@ def auth_update_viewed(data, email, cursor, conn):
         nStored = cursor.fetchone()[0]
     except TypeError:
         return {'msg' : 'An error occured when grabbing users recently viewed recipes'}, 500
+    except ProgrammingError:
+        nStored = 0
 
-    # Calculate number of items to delete
-    nDelete = max(0, len(recipes) + nStored - RVSTORAGE)
-    # Delete n recentlyViewed from the database
-    cursor.execute("DELETE FROM user_recentlyviewed WHERE u_id=%s LIMIT %s;", (u_id, nDelete))
+    if isinstance(recipes, list):
+        # Calculate number of items to delete
+        nDelete = max(0, len(recipes) + nStored - RVSTORAGE)
+        # Delete n recentlyViewed from the database
+        if nDelete > 0:
+            cursor.execute("DELETE FROM user_recentlyviewed WHERE u_id=%s LIMIT %s", (u_id, nDelete))
 
-    # Insert recently viewed recipes into the db
-    for recipe in recipes:
-        try:
-            r_id = recipe['r_id']
-        except KeyError: # Invalid recipe entry. Ignore it
-            pass
-        cursor.execute("INSERT INTO user_recentlyviewed(u_id, r_id) VALUES (%s, %s)", (u_id, r_id))
+        # Insert recently viewed recipes into the db
+        for recipe in recipes:
+            try:
+                r_id = recipe['r_id']
+            except KeyError: # Invalid recipe entry. Ignore it
+                pass
+            cursor.execute("INSERT INTO user_recentlyviewed(u_id, r_id) VALUES (%s, %s)", (u_id, r_id))
 
     conn.commit()
     return {'msg' : 'Successfully updated recently viewed table'}, 200
